feat: add searchbar to header. refactor header into Content component
